fetch/metadata/generated/script-module-import-dynamic.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/script-module-import-dynamic.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for dynamic ECMAScript module import</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-origin destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpSameSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpCrossSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy cross-site destination');
  </script>
</html>
